requirements.txt
dwave-ocean-sdk
dwave-samplers
networkx
numpy
pandas
matplotlib
seaborn
jupyter
scipy
